Move web correlation utilities to core
