+++ core/build.gradle
@@ -114,6 +114,7 @@ dependencies {
     compile ([group: 'commons-io', name: 'commons-io', version: '2.6' ])
     compile ([group: 'org.apache.commons', name: 'commons-lang3', version: '3.7'])
     compile ([group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.5.3'])
+    compile ([group: 'org.apache.httpcomponents', name: 'httpasyncclient', version: '4.1.3'])
     compile ([group: 'com.google.code.gson', name: 'gson', version: '2.8.2'])
     compile ([group: 'com.google.protobuf', name:'protobuf-java', version:'3.6.1'])
     compile ([group: 'io.grpc', name: 'grpc-stub', version: '1.16.1'])
