+++ core/src/main/java/com/microsoft/applicationinsights/web/internal/correlation/AppProfileFetcher.java
