+++ core/src/main/java/com/microsoft/applicationinsights/web/internal/correlation/ProfileFetcherResult.java
