+++ core/src/test/java/com/microsoft/applicationinsights/web/internal/correlation/mocks/MockHttpEntity.java
